Only add a space if variable declaration needs a type string.  Fix spelling of "traversal"
In Python, variables need no type declaration.  We can specify an empty string for
the declaration, but if the generator adds a space (between the "empty" string and
the variable name), it breaks Python spacing.  Add a check to only add the space if
the type declaration string is non-empty.
Correct spelling of traversal, both in comments and in operational code.