Improve comments. Sort imports for consistency. Use new var; don't change existing one.
Improve the comments about the steps in the generation process.

Sometimes C# imports, e.g.,
                using System.Text.RegularExpressions;
                using System.IO;
are produced in different order.  Sort the imports to try to have them always in a
consistent order.  Here is the piece of code:
    format(import_content) for import_content in sorted(set(self.filtering.imports))
                                                 ^^^^^^
In one place a variable is given a completely different value.  Use a new variable
instead of reassigning an existing variable, to reduce the chance that a future
programmer gets confused.  Here are the old and new assignments:
-        imports_content = self.file_template.generate_imports(imports_content)
+        imports_code = self.file_template.generate_imports(imports_content)