Remove the effect of --debug.  (Update UI for not changing flaw_groups.)
Previously, --debug reported when a file was created and it prevented writing the
license.  I left the code as a convenient hook if programmers want to use it.

Update the UI in test_suite_gen to say that flaw_groups must be given as coded.  VTSG
no longer "mangle" the names.