Factor out function to create the directory flaw group.
The code used to prepend OWASP to the name of the flaw group ("fixed" by 8 Feb 2022
commit eef8a585ad534e8cf...).  This was done in *two* places in code.  I fixed it in
one place, but not in the other.  So manifest files were placed in parallel
directories, with non-messed-up names.
    Since the same function computes the directory in both cases, such differences
cannot occur again.  Of course, now the "computation" is only concatenation, since
error would be unlikely to occur.  However, now it is (more) clear that both should
have the same value.