For PHP move {{flaw}} to its proper places in sink code.
The {{flaw}} line was in file_template.xml, where it does no good.  Remove it.  Add a
{{flaw}} line to each sink <code>, so they appear in the final generated code.

I found this error while fixing the bug in find_flaw().  Fixing the bug caused all
the <flaw ... lines to disappear from the manifest, since no "//flaw" lines were
found.  In researching what was going on and what the right thing should be, I
discovered this error.