Add test cases for misuse of INDENT ... ENDINDENT.
Note: the code that this tests has NOT been committed, yet.
This also commits some updates to Makefile: remove commented-out bit of script to
execute Python programs: that is now in gen_and_check and works.  Also fix built-in
self-test for test020 and update number of tests generated by Python.