Rename function in sample.py to be descriptive AND be different from another function.
Rename generate_file_name() in sample.py to be module_description().  First, the
function returns the description of the module; it does not generate the file name
(although the description is used in the name of the file). Second, there is another
function named generate_file_name() that DOES generate the file name.

Also fix and improve comments that I ran across in input_sample.py and generator.py.