Tweak: use Vulnerable instead of Vulnerability in Python tests. Don't execute class cases.
Change gen_and_check to NOT execute Python test cases that use classes.  Currently
those cases are not valid (the case name is wrong).
The phrase "Vulnerable: user command run" seems better than "Vulnerability: ..."
Change the hardcoded input string and the input supplied to be the former.  Yes, it's
a tiny change that doesn't affect functionality and likely nobody will ever notice,
but I wanted to do it.