Complete reference to test suite directory structure.  Fix overfull lines.  Misc. changes.
There was a note in Sec 5 about the directory structure of the generated test suite.
Since that was already described earlier, reference that description.  Remove the
example file name: it only described the complexities and a complete example is given
in the next section (refer to that).  Fix the worst of the overfull lines.  Many
other small improvements.