Handle missing flaw_group as empty flaw group.  Add test for missing flaw_type.
Add code in sink_sample.py so that a missing flaw_group in a sink flaw_type is an
empty flaw group.  Change the report to say "Empty group ..." instead of " group ..."
(a leading blank).  In the manual explain that a missing or empty flaw_group causes
the flaw type subdirectory to be placed directly under the language directory.  Add
test for missing flaw_group in test010.
    Add test for empty flaw_type, test004.  Add it to testVarious.