Make sure cases are not overwritten by duplicate names.
If the same <dir>s occur in different combinations of modules, the same case file
name will be generated and the file would be overwritten.  Add code to make sure
(previously generated) files are not overwritten.  Add test002 as an example.

Also change internal language file names to be more consistent, e.g., "sinks" instead
of "sink" and "filters" instead of "filtering".

Oops, I do NOT intend to include the SARIF writer.  It is not even hooked in to
VTSG.