Add .png files used in the manual.
Although fig_example_code1.png and fig_example_code2.png are derived from the .tif
originals, it requires an uncommon package to create them.  Instead of insisting that
everyone that wants to create the manual have that package, just add the derived .png
files.  Also, remove them from .gitignore and add rules in the Makefile to (try to)
recreate them if the sources change.