Make unbalanced INDENT ... DEDENT "user level" errors in src/synthesize.py.
Since the user can (easily) have unbalanced INDENT ... DEDENT lines in code, catch
and report them by regular checks and error reports.  They used to be assertions,
which are for *code* errors: something the user should not be able to trigger (unless
there are bugs in the code).  This led to changing the built-in checks in Makefile
and adding expected outputs for test016 and test017.