Add CWE 209 to Python.  Also add Juliet complexities and conditions.
Add an example CWE 209 from Juliet to Python.  Improve gen_and_check to
print a big error message if it finds a new sink and exits and add a (null) attack
for the new sink.  Add Juliet complexities 2 u & v, 3 u & v, j16 e & n.  Add 13 and
14 (switch statements where the case value DIFFERS from the match value) and 20n
(where it is NOT executed).  Update the Makefile for the new number of test cases.