Add file and env var inputs to path traversal in Python.
These inputs are actually filters, so that they will be wrapped in complexities.
Improve gen_and_check to create a file with the attack when the input comes from a
file.  Rename path_traversal to the more specific rel_path_trav(ersal).
Add an assert to the array complexity that the incoming variable is NOT None.  The
array snippet uses None to skip some things.  If the incoming variable is None, the
code is incorrectly NOT executed.  It would be nice to just fix the array code so it
always works, but I can't think of a way to do it.  With the assert, it now aborts
with an exception instead of silently NOT exercising the vulnerability.