Improve remove_indents() code and add comments.  Add an automated test, too.
The built-in self-test is not very useful: it compares one implemention of
remove_indents() with another.  But I put in so much work building the test
generator, I didn't want to throw it away.  It's not external, like those under
tests, because the routine is hard to access, except by running vtsg.