Add CWE15 External Control of System or Configuration Setting
Add CWE15 as a new sink in Python (py).  Change the "filters" (inputs) added for
CWE23 to be available for ALL flaws (not just CWE23).
In the tests, add the new files expected to be generated and update
manifest.xml.  Update the exec and photo files, too.  Add a new line in
gen_and_check to recognize this sink.
Add (back) a line of code in remove_indents() in case it is called with All
(which it never is in vtsg).