Add Kratkiewicz cases to Python.  Warn if a filter lacks {{out_var_name}}.  Other minor.
This doesn't generate ALL Kratkiewicz cases.  Many are just control flow
elaborations, so I coded mostly the most complicated ones.  I didn't try to code and
of the do...while cases or most of the irregular for loop cases.  Most additions are
in the sinks, with two inputs: one for ok array index (9) and one for the minimum
violation (10).  There are commented-out inputs for medium (16) and large (4096)
violations, in case someone wants to be more complete.  Add two filters: one for ok
(correct) checking of input index and one for minimally incorrect checking.  The
comment for each sink is annotated with the name of the Kratkiewicz case in the SARD.
    Add new expected test cases and update expected outputs.

A filter must assign a value to {{out_var_name} in every possible execution.  This
warning is an approximation, since the {{out_var_name}} could be in a comment or only
be assigned in some execution paths.

Other minor changes:
Make module_description a property instead of a method.  It makes for slightly
cleaner code, and all the other access functions are properties.  (This may have been
a method because it runs a bit of code instead of just returning a value.)

Add a case in test010 that uses exec_query, so that testAll exercises using them.
Update expected results.

Change TestCLI* cases to start with lower case (testCLI*), like all the other test
cases.  This was in Makefile and renaming some expected output files.