Move options need_complexity from sinks to filters.
In many cases need_complexity wasn't even needed in the C# sinks where it was used.
That flag does pretty much the same thing as the indirect don't-wrap-filter-in-
complexities-if-input_type-nofilter.  Putting it in filters, instead of sinks, gives
much more direct control of VTSG behavior---less need for magic.  So I replaced the
don't-wrap-filter-... code with the explicit options need_complexity control.  I had
to make small changes in several sink and filter files.  Update manual, too.

Improve some user error messages to give the name of the module with the problem,
instead of just saying "sinks file" or "filters file".