Compute if filter is executed in TestCase, not ComplexitiesGenerator
This allows that boolean to be entirely local (private) to TestCase and simplifies
its computation to be three simple lines instead of being intertwined with a BUNCH of
other complicated code.  It also allows us to move the call to ComplexitiesGenerator
to be with the rest of the compose code, instead of in TestCase __init__().  Moving
the call to ComplexitiesGenerator to be with the rest of the compose code allows us
to remove another TestCase-global variable (complexity_generator): it is now local to
the compose function.