Summarize cases selected.  Fix typos and improve comments.
In addition to the report of generated cases, accumulate stats for cases selected.  A
line to *print* the selected cases summary is commented out.
Change need_condition() to handle_condition() to be different from a function of the
same name in complexity.py and to make it more descriptive.
Improve some comments and fix some typos.