Move instuctions for mono and mcs to C# appendix.  Misc. fixes and improvements.
It is better for the instructions for installing mono and mcs to be in the C#
appendix because not everyone that runs VTSG needs them.  Add hints about how to
use homebrew to install them on MacOS.
Fix explainations of why Filters P and Q are not used, currently Fig. 12,
page 33, Sec. 5.1.  The reasons were the reverse of the figure.
Add a column that explains what each value of "executed" means for the example
code, currently Table 4, page 28, Sec. 4.8.2.  Improve spacing a bit.
Other minor improvements and typo fixes.