Capture stderr from test025su to check the error message.
At some point docopt is catching if the user gives both -s and -u at the same
time and producing a warning.  However docopt gives the warning on stderr,
whereas vtsg.py gave the warning on stdout.  So test025su_photo was empty.
Change the Makefile rule to capture stderr from the run and update the expected
photo file.