Add target to (re)generate ALL test025 cases.  Simplify rules for safe/unsafe tests.
The test025 target is NOT triggered during testing.  Running the safe and
running the unsafe cases should be equivalent.  I added the target in case I
want to rerun or recreate all the cases.

The rules for test025u and test025s had "pwd" then a grep -v to get rid of that
line!  Remove both commands to get the same result.