Document choosing cases with ACTS.  Misc other text improvements.
Add to the command line options table, and add a section that explains what -c
ACTS does in a little detail.  Add some text about the change in VTSG
architecture: save generated cases in a list, (optionally) choose cases, then
compose and write cases.  (VTSG used to compose and write each case as soon as
it was generated.)
   Make other miscellaneous text improvements that I found while doing the above.