Remove accomplished items.  They deal with C# and produce % of cases.
Sonika Sharma made sure C# code runs, so remove that.  We have a bash script to
compile and run C# code.  Note: the script does NOT check that the code executes
as intended, nor does the script execute a Proof of Vulnerability (PoV).
    We changed -n to accomplish the goal of produce p% of cases, so remove those
paragraphs.  I don't think pseudorandomly selecting cases to write would be
useful.  It is easy enough to add if someone thinks of a reason for it.