Let complexities and conditions <import> code, too.
Factor out code to get <imports> into a new function, get_imports(), in sample.py.
Change all existing code to use it.  Invoke it for complexities and conditions, too.
Add import attributes to complexities and conditions as well as accessors.  (For
future work, a stub condition object (well, the attributes) is in the complexity
class so that generation knows which complexity to pair it with.
    Add a new test langauge, test027, to check imports.  It exercises new features
like having two imports in a condition.  Note: this language is also built to test
Python classes.  Since VTSG doesn't support that now, there are Python exceptions
when we run those.  Add rules in the Makefile to test it.  These rules are for
development, too, so I will remove or modify some of them once all the class
importing works.
    Update the manual for imports in conditions and complexities.

To Do: change the Python, PHP, and C# language files to have imports in conditions,
instead of putting everything that might be needed in file_template.xml.  Update DTD
where needed.