Add filter to remove bad characters - CWE78. Fix hardcode input to NOT be safe for CWE78.
This filter uses regular expressions to remove bad characters.
The hardcoded value in that particular input IS an attack on CWE78: command line
injection.  Therefore using that is NOT safe UNLESS it is filtered.
    Add hundreds of new CWE78 cases.  Update Python_photo and Python_exec files.
Also update the OWASP_a1 manifest.  I did NOT check that the (new) manifest entries
are correct.