Make getXML() a regular function and change name to get_XML_file().
It was a class method of FileManager, but didn't use anything from or provide
anything to that class.  It is cleaner for it to simply be a function.  Even though
it is only called from generator.py (and therefore might be moved there), it needs
the os library, which generator.py does not import.