Improve hardcode to import any files. Internal change: generate file names earlier.
VTSG creates file names with minus signs.  Because of the grammar, Python cannot
import files if the name contains a minus sign.  The simpliest approach seems to be
to open code statements that import handle any string.  This change is to generallize
the code a bit to use os.path.join() instead of catenating strings.  This changes
many expected cases.
    I changed some code to move file name generation earlier.  This is a first step
to importing the appropriate file, not just the first extra file.