Handle import class correclty in class codes.  Improve test027 complexity.
This doesn't fully handle Python classes, but it is an important big step.

In the code that will be written to auxiliary files, e.g. class files in
classes_code, macro expand {{body_file}} entries (to have the name of the file that
will be imported) and splice in Python code that imports files with *any* name.
  Get the file name to be imported from classes_code.
  Factor out the code that splices in Python code that imports files with any name
into a new function splice_python_import().
  Factor out the code that actually writes code to one file into a new function
write_one_file().

In test language test027, instead of importing module_1 for both complexities 70bi
and 70ci, import a complexity-specific name, module70bi or module70ci, to make the
test slightly broader.  Change the expected generated files accordingly.

Improve test027 rules in Makefile to capture stderr ouput and to execute cases ending
in *i.py, too.