Add code to add assignments to initialize variables *only when needed.*
A filter has the following data flow:
    {{out_var_name}} = some_processing( {{in_var_name}} )
But if the filter is guarded (specifically, may not be executed):
    if rand() < 3:           # complexity & condition
        {{out_var_name}} = some_processing( {{in_var_name}} )   # filter
then out_var_name is not initialized.  (This doesn't arise much in C# or PHP
because all variables are initialized when they are declared.
    An assignment statement before the filter block properly initializes the out var
in case the filter is not executed:
    {{out_var_name}} = {{in_var_name}}   # synthesized assignment
    if rand() < 3:           # complexity & condition
        {{out_var_name}} = some_processing( {{in_var_name}} )   # filter
Previously VTSG ALWAYS added such an assignment after the input block.  This is not
needed in some cases. This change makes sure assignments are added, and
initializations are done, whenever they are needed and ONLY when they are needed.
This improvement changes lots of the C# and Python expected test cases.

In the manual document the new {{body_file}} meta variable.  Note that there is
Python-specific code in VTSG to handle importing files with names that VTSG chose.

In the source code, remember if the "filter" (or filter block, if it is wrapped in
complexities) will be executed.  Add an initializing assignment if and only if the
inner filter block is not always executed.

Update expected cases files for C#, Python, and test "languages", test023, test025,
and test027.  Update Some expected self-test outputs, too.

Remove the 6 Sep 2022 task to support classes in Python.  Add two more ideas, and
thoughts to some ideas, and a few misc. improvements.