Improve compilationTester.sh a little.
Note that this compiles the cases AND runs them.  Supply a command line argument and
standard input to the execution so more cases actually do something interesting.
Don't execute EQ_sql_server cases because they take about 20 seconds to time out;
with thousands of cases, those add HOURS to time to execute for little benefit.
Remove bits of code that skipped old style "second" files, that is, File2.  Those are
not generated anymore.