Fix errors in PHP language files.  Add all 285,123 expected PHP cases.  Improve run_PHP.
Remove mysql_real_escape_string() filter; it was removed because it was not secure.
Reduce the expected number of generated cases accordingly.  Add <import> to class
complexities so that they import the class (*b.php) files.  Look for tainted.SH as a
script to execute, not tainted.PHP; the script was a shell script, not a PHP script.
In run_PHP.sh, skip POST and SESSION input cases, since I don't know how to supply
either of those to php-cgi.  Simplify the coding that write PASSED in green or FAILED
in red.

Note: there are still problems in some of the PHP cases, but I don't have time to
figure them out and come up with a fix.